+++ SoftLayer/managers/metadata.py
@@ -6,7 +6,7 @@
     :copyright: (c) 2013, SoftLayer Technologies, Inc. All rights reserved.
     :license: BSD, see LICENSE for more details.
 """
-from SoftLayer.transport import make_rest_api_call
+from SoftLayer.transports import make_rest_api_call
 from SoftLayer.consts import API_PRIVATE_ENDPOINT_REST, USER_AGENT
 from SoftLayer.exceptions import SoftLayerAPIError, SoftLayerError
 
@@ -41,6 +41,8 @@ class MetadataManager(object):
 
         Usage:
 
+            >>> import SoftLayer
+            >>> client = SoftLayer.Client()
             >>> from SoftLayer.metadata import MetadataManager
             >>> meta = MetadataManager(client)
             >>> meta.get('datacenter')
